// client/src/app/shared/models/order.ts
export interface Order {
    id: number
    orderDate: string
    buyerEmail: string
    shippingAddress: ShippingAddress
    deliveryMethod: string
    shippingPrice: number
    paymentSummary: PaymentSummary
    orderItems: OrderItem[]
    subTotal: number
    total: number
    status: string
    paymentIntentId: string
}

export interface ShippingAddress {
    name: string
    line1: string
    line2?: string
    city: string
    state: string
    postalCode: string
    country: string
}

export interface PaymentSummary {
    last: number
    brand: string
    expMonth: number
    year: number
}

export interface OrderItem {
    productId: number
    productName: string
    pictureUrl: string
    price: number
    quantity: number
}

export interface OrderToCreate{
    cartId: string;
    deliveryMethodId: number;
    shippingAddress: ShippingAddress;
    paymentSummary: PaymentSummary;

}